+++ src/ideations/ideations.controller.ts
@@ -13,7 +13,7 @@ import {
 import { IdeationsService } from "./ideations.service";
 import { CreateIdeationDto } from "./dto/create-ideation.dto";
 import { UpdateIdeationDto } from "./dto/update-ideation.dto";
-import { ApiOperation, ApiResponse, ApiTags } from "@nestjs/swagger";
+import { ApiOperation, ApiResponse, ApiTags, ApiParam } from "@nestjs/swagger";
 import {
     BadRequestErrorResponse,
     ConflictErrorResponse,
@@ -28,7 +28,13 @@ import {
 import { AppPermissions } from "../auth/auth.permissions";
 import { Permissions } from "../global/decorators/permissions.decorator";
 import { CustomRequest } from "../global/types/CustomRequest";
+import { Roles } from "../global/decorators/roles.decorator";
+import { AppRoles } from "../auth/auth.roles";
 
+//
+//OWN_TEAM permission requires a :teamId param, but does not correctly enforce same team rule
+//:ideationId param is not currently used to check if user has permission to change it
+//
 @Controller()
 @ApiTags("Voyage - Ideations")
 export class IdeationsController {
@@ -236,5 +242,92 @@ export class IdeationsController {
             teamId,
             ideationId,
         );
+    }
+
+    @ApiOperation({
+        summary: "Selects one ideation as team project for voyage.",
+    })
+    @ApiResponse({
+        status: HttpStatus.UNAUTHORIZED,
+        description: "User is not authorized to perform this action.",
+        type: UnauthorizedErrorResponse,
+    })
+    @ApiResponse({
+        status: HttpStatus.NOT_FOUND,
+        description: "Ideation with given ID does not exist.",
+        type: NotFoundErrorResponse,
+    })
+    @ApiResponse({
+        status: HttpStatus.CONFLICT,
+        description: "An ideation has already been selected.",
+        type: ConflictErrorResponse,
+    })
+    @ApiResponse({
+        status: HttpStatus.CREATED,
+        description: "Successfully selected ideation.",
+        type: IdeationResponse,
+    })
+    @ApiParam({
+        name: "teamId",
+        description: "voyage team Id",
+        type: "Integer",
+        required: true,
+        example: 1,
+    })
+    @ApiParam({
+        name: "ideationId",
+        description: "ideation Id",
+        type: "Integer",
+        required: true,
+        example: 1,
+    })
+    @Permissions(AppPermissions.OWN_TEAM)
+    @Post("/:ideationId/select")
+    setIdeationSelection(
+        @Request() req: CustomRequest,
+        @Param("teamId", ParseIntPipe) teamId: number,
+        @Param("ideationId", ParseIntPipe) ideationId: number,
+    ) {
+        return this.ideationsService.setIdeationSelection(
+            req,
+            teamId,
+            ideationId,
+        );
+    }
+
+    @ApiOperation({
+        summary: "Clears the current ideation selection for team.",
+        description: "Admin only allowed.",
+    })
+    @ApiResponse({
+        status: HttpStatus.UNAUTHORIZED,
+        description: "User is not authorized to perform this action.",
+        type: UnauthorizedErrorResponse,
+    })
+    @ApiResponse({
+        status: HttpStatus.NOT_FOUND,
+        description: "Ideation for this team does not exist.",
+        type: NotFoundErrorResponse,
+    })
+    @ApiResponse({
+        status: HttpStatus.CREATED,
+        description: "Successfully cleared ideation selection.",
+        type: IdeationResponse,
+    })
+    @ApiParam({
+        name: "teamId",
+        description: "voyage team Id",
+        type: "Integer",
+        required: true,
+        example: 1,
+    })
+    @Permissions(AppPermissions.OWN_TEAM)
+    @Roles(AppRoles.Admin)
+    @Post("/reset-selection")
+    resetIdeationSelection(
+        @Request() req: CustomRequest,
+        @Param("teamId", ParseIntPipe) teamId: number,
+    ) {
+        return this.ideationsService.resetIdeationSelection(req, teamId);
     }
 }