+++ src/sprints/sprints.response.ts
@@ -18,6 +18,20 @@ class Sprint {
     teamMeetings: number;
 }
 
+class SprintWithoutMeetings {
+    @ApiProperty({ example: 1 })
+    id: number;
+
+    @ApiProperty({ example: 1 })
+    number: number;
+
+    @ApiProperty({ example: "2024-01-08T00:00:00.000Z" })
+    startDate: Date;
+
+    @ApiProperty({ example: "2024-01-14T00:00:00.000Z" })
+    endDate: Date;
+}
+
 export class VoyageResponse {
     @ApiProperty({ example: 1 })
     id: number;
@@ -45,6 +59,33 @@ export class VoyageResponse {
     sprints: Sprint;
 }
 
+export class VoyageResponseWithoutMeetings {
+    @ApiProperty({ example: 1 })
+    id: number;
+
+    @ApiProperty({ example: "47" })
+    number: string;
+
+    @ApiProperty({ example: "2023-12-31T00:00:00.000Z" })
+    soloProjectDeadline: Date;
+
+    @ApiProperty({ example: "2024-02-25T00:00:00.000Z" })
+    certificateIssueDate: Date;
+
+    @ApiProperty({ example: null })
+    @Optional()
+    showcasePublishDate: Date;
+
+    @ApiProperty({ example: "2024-01-08T00:00:00.000Z" })
+    startDate: Date;
+
+    @ApiProperty({ example: "2024-02-18T00:00:00.000Z" })
+    endDate: Date;
+
+    @ApiProperty({ isArray: true })
+    sprints: SprintWithoutMeetings;
+}
+
 class Agenda {
     @ApiProperty({ example: 1 })
     id: number;
@@ -62,6 +103,11 @@ class Agenda {
     status: boolean;
 }
 
+class AgendaMeetingResponse extends Agenda {
+    @ApiProperty({ example: "2023-11-30T06:47:11.694Z" })
+    updatedAt: Date;
+}
+
 export class MeetingResponseWithSprintAndAgenda {
     @ApiProperty({ example: 1 })
     id: number;
@@ -71,6 +117,12 @@ export class MeetingResponseWithSprintAndAgenda {
 
     @ApiProperty({ example: "First sprint kickoff meeting" })
     title: string;
+
+    @ApiProperty({
+        example:
+            "Lorem ipsum dolor sit amet consectetur adipiscing elit magna praesent, nunc metus egestas nam libero quisque senectus facilisis, dis nec gravida sodales sagittis duis risus parturient. Eu scelerisque gravida posuere blandit interdum iaculis venenatis rhoncus taciti, tempus nullam cras eros quisque himenaeos condimentum auctor cursus, leo neque montes mollis litora imperdiet luctus purus.",
+    })
+    description: string;
 
     @Optional()
     @ApiProperty({ example: "2024-01-08T00:00:00.000Z" })
@@ -85,7 +137,7 @@ export class MeetingResponseWithSprintAndAgenda {
     notes: string;
 
     @ApiProperty({ isArray: true })
-    agendas: Agenda;
+    agendas: AgendaMeetingResponse;
 }
 
 export class MeetingResponse {
@@ -97,6 +149,12 @@ export class MeetingResponse {
 
     @ApiProperty({ example: "Sprint Planning" })
     title: string;
+
+    @ApiProperty({
+        example:
+            "Lorem ipsum dolor sit amet consectetur adipiscing elit magna praesent, nunc metus egestas nam libero quisque senectus facilisis, dis nec gravida sodales sagittis duis risus parturient. Eu scelerisque gravida posuere blandit interdum iaculis venenatis rhoncus taciti, tempus nullam cras eros quisque himenaeos condimentum auctor cursus, leo neque montes mollis litora imperdiet luctus purus.",
+    })
+    description: string;
 
     @Optional()
     @ApiProperty({ example: "2023-12-02T13:37:51.465Z" })