+++ src/sprints/sprints.service.ts
@@ -90,11 +90,6 @@ export class SprintsService {
                         number: true,
                         startDate: true,
                         endDate: true,
-                        teamMeetings: {
-                            select: {
-                                id: true,
-                            },
-                        },
                     },
                 },
             },
@@ -152,6 +147,7 @@ export class SprintsService {
                     },
                 },
                 title: true,
+                description: true,
                 dateTime: true,
                 meetingLink: true,
                 notes: true,
@@ -161,6 +157,7 @@ export class SprintsService {
                         title: true,
                         description: true,
                         status: true,
+                        updatedAt: true,
                     },
                 },
                 formResponseMeeting: {
@@ -211,7 +208,13 @@ export class SprintsService {
     async createTeamMeeting(
         teamId: number,
         sprintNumber: number,
-        { title, meetingLink, dateTime, notes }: CreateTeamMeetingDto,
+        {
+            title,
+            description,
+            meetingLink,
+            dateTime,
+            notes,
+        }: CreateTeamMeetingDto,
     ) {
         const sprintId = await this.findSprintIdBySprintNumber(
             teamId,
@@ -242,6 +245,7 @@ export class SprintsService {
                 sprintId,
                 voyageTeamId: teamId,
                 title,
+                description,
                 meetingLink,
                 dateTime,
                 notes,
@@ -251,7 +255,13 @@ export class SprintsService {
 
     async updateTeamMeeting(
         meetingId: number,
-        { title, meetingLink, dateTime, notes }: UpdateTeamMeetingDto,
+        {
+            title,
+            description,
+            meetingLink,
+            dateTime,
+            notes,
+        }: UpdateTeamMeetingDto,
     ) {
         try {
             const updatedMeeting = await this.prisma.teamMeeting.update({
@@ -260,6 +270,7 @@ export class SprintsService {
                 },
                 data: {
                     title,
+                    description,
                     meetingLink,
                     dateTime,
                     notes,