+++ src/techs/techs.response.ts
@@ -54,12 +54,12 @@ export class TeamTechResponse {
 
 export class TechItemResponse {
     @ApiProperty({ example: 10 })
-    id: number;
+    teamTechStackItemVotedId: number;
 
     @ApiProperty({ example: 11 })
     teamTechId: number;
 
-    @ApiProperty({ example: "Frontend" })
+    @ApiProperty({ example: 8 })
     teamMemberId: number;
 
     @ApiProperty({ example: "2023-12-01T13:55:00.611Z" })
@@ -67,4 +67,12 @@ export class TechItemResponse {
 
     @ApiProperty({ example: "2023-12-01T13:55:00.611Z" })
     updatedAt: Date;
+}
+
+export class TechItemDeleteResponse {
+    @ApiProperty({ example: "The vote and tech stack item were deleted" })
+    message: string;
+
+    @ApiProperty({ example: 200 })
+    statusCode: number;
 }