+++ test/techs.e2e-spec.ts
@@ -126,7 +126,7 @@ describe("Techs Controller (e2e)", () => {
                 .expect((res) => {
                     expect(res.body).toEqual(
                         expect.objectContaining({
-                            id: expect.any(Number),
+                            teamTechStackItemVoteId: expect.any(Number),
                             teamTechId: expect.any(Number),
                             teamMemberId: expect.any(Number),
                             createdAt: expect.any(String),
@@ -228,7 +228,7 @@ describe("Techs Controller (e2e)", () => {
                 .expect((res) => {
                     expect(res.body).toEqual(
                         expect.objectContaining({
-                            id: expect.any(Number),
+                            teamTechStackItemVoteId: expect.any(Number),
                             teamTechId: expect.any(Number),
                             teamMemberId: expect.any(Number),
                             createdAt: expect.any(String),
@@ -321,11 +321,8 @@ describe("Techs Controller (e2e)", () => {
                 .expect((res) => {
                     expect(res.body).toEqual(
                         expect.objectContaining({
-                            id: expect.any(Number),
-                            teamTechId: expect.any(Number),
-                            teamMemberId: expect.any(Number),
-                            createdAt: expect.any(String),
-                            updatedAt: expect.any(String),
+                            message: "This vote was deleted",
+                            statusCode: 200,
                         }),
                     );
                 });
@@ -339,6 +336,34 @@ describe("Techs Controller (e2e)", () => {
                 },
             });
             return expect(techStackVote[0]).toEqual(undefined);
+        });
+
+        it("should return 200 if tech last vote was deleted and team tech stack item is deleted", async () => {
+            const teamId: number = 2;
+            const techId: number = 9;
+            return request(app.getHttpServer())
+                .delete(`/voyages/teams/${teamId}/techs/${techId}`)
+                .set("Authorization", `Bearer ${userAccessToken}`)
+                .expect(200)
+                .expect("Content-Type", /json/)
+                .expect((res) => {
+                    expect(res.body).toEqual(
+                        expect.objectContaining({
+                            message:
+                                "The vote and tech stack item were deleted",
+                            statusCode: 200,
+                        }),
+                    );
+                });
+        });
+
+        it("- verify that tech stack Item was deleted from database", async () => {
+            const techStackItem = await prisma.teamTechStackItem.findFirst({
+                where: {
+                    name: newTechName,
+                },
+            });
+            return expect(techStackItem).toBeNull();
         });
 
         it("should return 401 unauthorized if not logged in", async () => {